Add Journal support, improve loading performance
This commit re-enables full indexing of all events, which uses a lot of RAM spikes CPU for a short time.

That will likely change again but has been useful for optimizing; as have additional metrics I've added to the launcher.

Implementing a view of recent logs has also been useful for performance improvements.

This is work in progress (but fully working) implmentation that I'm going to refactor.

Event names, variables and contentions need a bit of a cleanup.