Add icons to app and refactoring toolbar
I've chosen to manually include fonts in the web app so that work on new font assets can be done in parallel to app development, so that I'm not inadventantly making breaking changes in the app when there are changes to the icon font.

This also works well as a flow if the font icon ends up moving to it's own repository.