Get system map plotter to output co-ordinate data
Looks like the plotter is working to plot the system bodies and stations in a useful way.

The next step will be to hook this up to a suitable socket event able to return data for the current system (instead of dummy data).

I may migrate this code to the backend service when I do that.

Both of those things should be pretty trivial.

After that comes the harder part of porting the SVG render to Next.js + React.