Fixes for fullscreen mode in Safari
Uses addtional webkit method checks for testing if is in full screen mode and for exiting full screen mode in Safari.

The HTML element needs a background colour to prevent a rendering glitch in Safari when switching to full screen, otherwise 2/3 of the screen renders with a white background.

It seems like Safari doesn't handle resizing elements or psuedo elements well when switching to fullscreen.