System map now rendering in colour!
I'm not actually sure why it's rendering in colour now as I actually expected it to render in monotone, but it's looking better.

There are some place holder textures included I need to replace (they are royalty free but need attribution and I'd have to dig out the sources - one is a NASA image IIRC).