Load last X days logs relative to most recent log
Loading relative to last active log is a  more intelligent mechanism as easier to ensure we have useful data to work with without importing what could be years worth of logs.

This does not mean we don't need persistance - there is still some useful data in old logs and there are still edge cases like players who play infreqeuently - and but helps defer the need for it.

Trying default of 30 days for now, may bump to 90 days depending on testing.