Refactor checkProcessAlreadyExists() to use mutex
The old approach of running Tasklist would fail on non-English language machines as it relied on string matching.

It now uses a mutex (which is much better approach) and should work reliabily.

Unlike a file lock it shouldn't also have the risk of being stale, even if the app crashes.