Refactor table animation to use observables
This solves the problem of animations being cut off on differently sized displays and also of animation delays when navigating to views with long lists.

I have decided that animations should only play when a view is first rendered and not when scrolled as the latter can mae the UI harder to read. Interestingly, this is also the approach the game mostly seems to take for similar UI elements.