Fix so that unknown blueprints don't cause errors
Things like the holiday decoration flak launchers are actually engineering blueprints applied to flak launchers, they are not in the list of blueprints I have and so were causing a problem.

The code is more defensive now and will not cause an error or fail to render the UI when it comes across a mod it doesn't recognize (although it will not show the mod in the UI if it doesn't have details for it).